refactor(docker): extract ContainerTable inline components into separate files
HostNameCell, ContainerSubTable, and ContainerNameCell were defined inline
in ContainerTable.tsx (~657 lines). Each is now in its own file under
components/docker/, reducing ContainerTable.tsx to focused orchestration logic.
