+++ src/components/docker/ContainerTable.tsx
@@ -1,9 +1,7 @@
-import { memo, useCallback, useEffect, useMemo, useRef, useState, type ReactNode } from 'react';
+import { useCallback, useMemo, useRef } from 'react';
 import type { ColumnDef, ExpandedState } from '@tanstack/react-table';
-import { Box, Chip, CircularProgress, IconButton, Typography } from '@mui/material';
-import { ChevronRight, History, Server, Settings, WifiOff } from 'lucide-react';
+import { Box, CircularProgress, Typography } from '@mui/material';
 import { useDockerSettings, useGeneralSettings } from '@/hooks/useSettings';
-import { useToast } from '@/hooks/toastAtom';
 import { StaleDataAlert } from '@/components/shared-table/StaleDataAlert';
 import { DataTable, type MetricGroup } from '@/components/shared-table/DataTable';
 import { metricColumn, nameColumn } from '@/components/shared-table/columns';
@@ -12,18 +10,16 @@ import { formatAsPercentParts, formatBytesParts, formatBitsSIUnitsParts } from '
 import type {
   DockerStatsRow,
   DockerStatsFromDB,
-  DockerHostTableRow,
   DockerContainerTableRow,
   DockerTableRow,
 } from '@/types/docker';
 import type { DockerInventorySnapshotContainer } from '@/types/docker-inventory';
 import { buildDockerTableHierarchy, rowToDockerStats } from '@/lib/utils/docker-hierarchy-builder';
-import { getIconUrl, FALLBACK_ICON_URL } from '@/lib/utils/icon-resolver';
-import IconPickerDialog from '@/components/docker/IconPickerDialog';
 import ContainerDetailPanel from '@/components/docker/ContainerDetailPanel';
-import ContainerStateChip from '@/components/docker/ContainerStateChip';
-import { usePulseIndicator } from '@/hooks/usePulseIndicator';
 import { buildContainerChartData } from '@/hooks/useContainerChartData';
+import HostNameCell from '@/components/docker/HostNameCell';
+import ContainerNameCell from '@/components/docker/ContainerNameCell';
+import ContainerSubTable from '@/components/docker/ContainerSubTable';
 
 export const DOCKER_ENTITY_ICONS_QUERY_KEY = ['docker-entity-icons'] as const;
 
@@ -435,223 +431,3 @@ export default function ContainerTable({
   );
 }
 
-/**
- * Nested DataTable for container rows within an expanded host.
- * Detail panel renders inline with MUI Collapse animation.
- */
-const ContainerSubTable = memo(function ContainerSubTable({
-  containers,
-  columns,
-  renderDetailPanel,
-  rowClassName,
-  rowAttributes,
-  isContainerExpanded,
-  toggleContainerExpanded,
-}: Readonly<{
-  containers: DockerContainerTableRow[];
-  columns: ColumnDef<DockerTableRow, unknown>[];
-  renderDetailPanel: (row: DockerTableRow) => ReactNode;
-  rowClassName: (row: DockerTableRow) => string;
-  rowAttributes: (row: DockerTableRow) => Record<`data-${string}` | `aria-${string}`, string>;
-  isContainerExpanded: (id: string) => boolean;
-  toggleContainerExpanded: (id: string) => void;
-}>) {
-  const containerExpanded = useMemo<ExpandedState>(() => {
-    const state: Record<string, boolean> = {};
-    for (const c of containers) {
-      state[c.id] = isContainerExpanded(c.id);
-    }
-    return state;
-  }, [containers, isContainerExpanded]);
-
-  const handleContainerExpandedChange = useCallback(
-    (newExpanded: ExpandedState) => {
-      if (typeof newExpanded === 'boolean') return;
-      const current = containerExpanded as Record<string, boolean>;
-      for (const [id, value] of Object.entries(newExpanded)) {
-        if (value !== (current[id] ?? false)) {
-          toggleContainerExpanded(id);
-        }
-      }
-      for (const id of Object.keys(current)) {
-        if (current[id] && !(id in newExpanded)) {
-          toggleContainerExpanded(id);
-        }
-      }
-    },
-    [containerExpanded, toggleContainerExpanded],
-  );
-
-  return (
-    <DataTable
-      data={containers}
-      columns={columns}
-      getRowId={(row) => row.id}
-      renderDetailPanel={renderDetailPanel}
-      expandedState={containerExpanded}
-      onExpandedChange={handleContainerExpandedChange}
-      rowClassName={rowClassName}
-      rowAttributes={rowAttributes}
-      enableSorting={false}
-      showHeader={false}
-    />
-  );
-});
-
-const HostNameCell = memo(function HostNameCell({
-  row,
-  expanded,
-}: Readonly<{
-  row: DockerHostTableRow;
-  expanded: boolean;
-}>) {
-  const { totalHosts, children, aggregated: a } = row;
-  const hasContainers = children.length > 0;
-  const canToggle = hasContainers && totalHosts > 1;
-
-  return (
-    <div className="flex items-center gap-2">
-      {canToggle && (
-        <ChevronRight
-          size={18}
-          className={`shrink-0 transition-transform duration-200 ${expanded ? 'rotate-90' : ''}`}
-        />
-      )}
-      <Server size={18} className="shrink-0" />
-      {row.isStale && (
-        <WifiOff size={16} className="text-amber-600 dark:text-amber-400 shrink-0" />
-      )}
-      <span className="font-bold">{row.hostName}</span>
-      {a.staleContainerCount > 0 && !row.isStale && (
-        <Chip size="small" variant="filled" color="warning" label={`${a.staleContainerCount} stale`} />
-      )}
-    </div>
-  );
-});
-
-
-/**
- * Name cell for container rows: shows icon, pulse indicator, state chip, name,
- * settings and history buttons. Uses hooks for pulse animation and icon picker.
- */
-function ContainerNameCell({
-  row,
-  expanded,
-  onIconChange,
-  onOpenHistory,
-}: Readonly<{
-  row: DockerContainerTableRow;
-  expanded: boolean;
-  onIconChange: (serviceKeyEntity: string, iconSlug: string) => Promise<void>;
-  onOpenHistory?: (containerId: string, host: string) => void;
-}>) {
-  const { inventory, stats } = row;
-
-  const { showToast } = useToast();
-  const [iconPickerOpen, setIconPickerOpen] = useState(false);
-  const [iconError, setIconError] = useState(false);
-
-  const iconUrl = getIconUrl(stats?.icon ?? null, inventory.image);
-
-  useEffect(() => {
-    setIconError(false);
-  }, [iconUrl]);
-
-  const lastChartRow = row.chartData?.at(-1);
-  const lastUpdated = lastChartRow ? new Date(lastChartRow.time) : undefined;
-  const lastUpdatedMs = lastUpdated?.getTime() ?? 0;
-  const { indicatorRef, pingRef, dotRef } = usePulseIndicator(lastUpdatedMs);
-
-  const handleIconSelect = async (iconSlug: string) => {
-    try {
-      const serviceKeyEntity = stats?.serviceKeyEntity ?? `${inventory.host}/${inventory.name}`;
-      await onIconChange(serviceKeyEntity, iconSlug);
-    } catch (err) {
-      console.error('Failed to update container icon:', err);
-      showToast('Failed to update icon. Please try again.', 'error');
-    }
-  };
-
-  const handleHistoryClick = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    const { host, containerId } = inventory;
-    if (host && containerId && onOpenHistory) onOpenHistory(containerId, host);
-  };
-
-  return (
-    <>
-      <div className="flex items-center gap-2">
-        <ChevronRight
-          size={16}
-          className={`shrink-0 transition-transform duration-200 ${expanded ? 'rotate-90' : ''}`}
-        />
-
-        {/* Pulse indicator: for running and restarting containers with data */}
-        {(inventory.state === 'running' || inventory.state === 'restarting') && (
-          <div
-            ref={indicatorRef}
-            className="relative w-2 h-2 shrink-0"
-            title={lastUpdated ? `Last updated: ${lastUpdated.toLocaleTimeString()}` : 'No data yet'}
-          >
-            <div
-              ref={pingRef}
-              className="absolute inset-0 rounded-full transition-opacity duration-200 opacity-0"
-              style={{ backgroundColor: 'var(--indicator-active)' }}
-            />
-            <div
-              ref={dotRef}
-              className="absolute inset-0 rounded-full transition-colors duration-300"
-              style={{ backgroundColor: 'var(--indicator-active)' }}
-            />
-          </div>
-        )}
-        {inventory.state !== 'running' && (
-          <ContainerStateChip state={inventory.state} />
-        )}
-
-        <img
-          src={iconError ? FALLBACK_ICON_URL : iconUrl}
-          alt=""
-          className="w-5 h-5 shrink-0"
-          onError={() => setIconError(true)}
-        />
-        <span className="truncate">{inventory.name}</span>
-        <IconButton
-          size="small"
-          onClick={(e) => {
-            e.stopPropagation();
-            setIconPickerOpen(true);
-          }}
-          className={`p-1! transition-opacity ${expanded ? 'opacity-100' : 'opacity-0 group-hover:opacity-100 focus-visible:opacity-100'}`}
-          aria-label="Change container icon"
-          tabIndex={expanded ? 0 : -1}
-          aria-hidden={!expanded}
-        >
-          <Settings size={14} />
-        </IconButton>
-        {onOpenHistory && (
-          <IconButton
-            size="small"
-            onClick={handleHistoryClick}
-            className={`p-1! transition-opacity ${expanded ? 'opacity-100' : 'opacity-0 group-hover:opacity-100 focus-visible:opacity-100'}`}
-            aria-label="View container history"
-            tabIndex={expanded ? 0 : -1}
-            aria-hidden={!expanded}
-          >
-            <History size={14} />
-          </IconButton>
-        )}
-      </div>
-
-      {iconPickerOpen && (
-        <IconPickerDialog
-          open={iconPickerOpen}
-          onClose={() => setIconPickerOpen(false)}
-          onSelect={handleIconSelect}
-          currentIcon={stats?.icon ?? null}
-          containerName={inventory.name}
-        />
-      )}
-    </>
-  );
-}
